+++ backend/spring/src/main/java/demetoir/vagi/config/auth/CustomOAuth2UserService.java
@@ -1,0 +1,27 @@
+package demetoir.vagi.config.auth;
+
+import lombok.RequiredArgsConstructor;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
+import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
+
+  private static final Log log = LogFactory.getLog(CustomOAuth2UserService.class);
+
+  @Override
+  public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
+    log.info("hserserser");
+
+    OAuth2UserService delegate = new DefaultOAuth2UserService();
+
+    return null;
+  }
+}
